backend/requirements.txt
fastapi>=0.134.0
uvicorn[standard]>=0.41.0
sqlmodel>=0.0.22
python-dotenv>=1.0.0
google-genai>=1.0.0
playwright>=1.40.0
apscheduler>=3.10.0
pypdf>=4.0.0
python-docx>=1.1.0
python-multipart>=0.0.22
httpx>=0.27.0
pytz>=2024.1
